requirements.txt
ultralytics
opencv-python
numpy
pandas
tqdm
shapely
pymupdf
pyyaml
openpyxl

